Fix for build failures on mac and Linux
BUG=none
TEST=it compiles
TBR=joi
Review URL: http://codereview.chromium.org/7327032

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@91978 0039d316-1c4b-4281-b951-d872f2087c98

+++ content/browser/renderer_host/resource_dispatcher_host.cc
@@ -1656,7 +1656,7 @@ void ResourceDispatcherHost::NotifyResponseStarted(net::URLRequest* request,
       BrowserThread::UI, FROM_HERE,
       NewRunnableFunction(
           &ResourceDispatcherHost::NotifyOnUI<ResourceRequestDetails>,
-          content::NOTIFICATION_RESOURCE_RESPONSE_STARTED,
+          static_cast<int>(content::NOTIFICATION_RESOURCE_RESPONSE_STARTED),
           render_process_id, render_view_id, detail));
 }
 
@@ -1674,7 +1674,7 @@ void ResourceDispatcherHost::NotifyReceivedRedirect(net::URLRequest* request,
       BrowserThread::UI, FROM_HERE,
       NewRunnableFunction(
           &ResourceDispatcherHost::NotifyOnUI<ResourceRedirectDetails>,
-          content::NOTIFICATION_RESOURCE_RECEIVED_REDIRECT,
+          static_cast<int>(content::NOTIFICATION_RESOURCE_RECEIVED_REDIRECT),
           render_process_id, render_view_id, detail));
 }
